64 bit requirement, comments

+++ proto/src/connection/fallible/mod.rs
@@ -2,6 +2,6 @@
 pub mod quic;
 pub mod tcp;
 
-/// Assert that we are at _least_ running on a 32-bit system
-/// TODO: find out if there is a better way than the `u32` cast
-const _: [(); 0 - (!(usize::BITS >= u32::BITS)) as usize] = [];
+/// Assert that we are at _least_ running on a 64-bit system
+/// TODO: find out if there is a better way than the `u64` cast
+const _: [(); 0 - (!(usize::BITS >= u64::BITS)) as usize] = [];
